backend-ragie/pyproject.toml
[tool.poetry]
name = "ragie-backend"
version = "1.0.0"
description = "Lean document management backend using Ragie.ai"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = ""}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
supabase = "^2.17.0"
ragie = "^1.15.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
stripe = "^11.4.0"
python-multipart = "^0.0.20"
pydantic = "^2.11.7"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
httpx = "^0.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
